Refactored postcss functionality in it's own file - #163
